Changed Base32hex/64 return data type to int32_t
refs #2137 @1h

+++ src/common/base32hex.h
@@ -46,7 +46,7 @@
  * \retval >=0		length of output string.
  * \retval -1		if error.
  */
-int64_t base32hex_encode(const uint8_t  *in,
+int32_t base32hex_encode(const uint8_t  *in,
 			 const uint32_t in_len,
 			 uint8_t        *out,
 			 const uint32_t out_len);
@@ -66,7 +66,7 @@ int64_t base32hex_encode(const uint8_t  *in,
  * \retval >=0		length of output string.
  * \retval -1		if error.
  */
-int64_t base32hex_encode_alloc(const uint8_t  *in,
+int32_t base32hex_encode_alloc(const uint8_t  *in,
 			       const uint32_t in_len,
 			       uint8_t        **out);
 
@@ -86,7 +86,7 @@ int64_t base32hex_encode_alloc(const uint8_t  *in,
  * \retval -1		if error.
  * \retval -2		if bad input data.
  */
-int64_t base32hex_decode(const uint8_t  *in,
+int32_t base32hex_decode(const uint8_t  *in,
 			 const uint32_t in_len,
 			 uint8_t        *out,
 			 const uint32_t out_len);
@@ -108,7 +108,7 @@ int64_t base32hex_decode(const uint8_t  *in,
  * \retval -1		if error.
  * \retval -2		if bad input data.
  */
-int64_t base32hex_decode_alloc(const uint8_t  *in,
+int32_t base32hex_decode_alloc(const uint8_t  *in,
 			       const uint32_t in_len,
 			       uint8_t        **out);
